+++ tests/testthat/test-init.R
@@ -21,7 +21,7 @@ test_that("we can initialize a project using 'breakfast'", {
   expected <- c("bread", "breakfast", "oatmeal", "toast")
   lockfile <- renv::snapshot(lockfile = NULL)
 
-  actual <- setdiff(names(renv_records(lockfile)), "renv")
+  actual <- setdiff(names(renv_lockfile_records(lockfile)), "renv")
   expect_setequal(actual, expected)
 
 })
@@ -36,7 +36,7 @@ test_that("we can initialize a project using 'toast'", {
   expected <- c("bread", "toast")
   lockfile <- renv::snapshot(lockfile = NULL)
 
-  actual <- setdiff(names(renv_records(lockfile)), "renv")
+  actual <- setdiff(names(renv_lockfile_records(lockfile)), "renv")
   expect_setequal(actual, expected)
 
 })