+++ tests/testthat/test-record.R
@@ -10,7 +10,7 @@ test_that("an existing lockfile can be updated", {
   renv::record(list(bread = "bread@0.1.0"))
 
   lockfile <- renv_lockfile_read("renv.lock")
-  records <- renv_records(lockfile)
+  records <- renv_lockfile_records(lockfile)
   expect_length(records, 1L)
   expect_equal(records$bread$Version, "0.1.0")
 
@@ -18,23 +18,23 @@ test_that("an existing lockfile can be updated", {
   renv::record(list(toast = "toast@1.0.0"))
 
   lockfile <- renv_lockfile_read("renv.lock")
-  records <- renv_records(lockfile)
+  records <- renv_lockfile_records(lockfile)
   expect_length(records, 2L)
   expect_equal(records$bread$Version, "0.1.0")
   expect_equal(records$toast$Version, "1.0.0")
 
   # use short-hand
   lockfile <- renv::record("toast@1.0.1", lockfile = lockfile)
-  records <- renv_records(lockfile)
+  records <- renv_lockfile_records(lockfile)
   expect_equal(records$toast$Version, "1.0.1")
 
   lockfile <- renv::record(list(toast = "1.0.2"), lockfile = lockfile)
-  records <- renv_records(lockfile)
+  records <- renv_lockfile_records(lockfile)
   expect_equal(records$toast$Version, "1.0.2")
 
   # remove a record
   lockfile <- renv::record(list(toast = NULL), lockfile = lockfile)
-  records <- renv_records(lockfile)
+  records <- renv_lockfile_records(lockfile)
   expect_true(is.null(records$toast))
 
 })